+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Add fixture check against deposit request hash in `DepositManagerBase` unit tests
 - Change Deposit events to include all info needed to recover deposit hash
 - Refactor unit tests to whitelist deposit source for deposits now that we use that instead of `msg.sender == deposit.spender`
 - Vault and BalanceManager now take both deposit and its `source` so they know where transfer funds from upon deposit