+++ packages/deposit-screener/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Unreleased
 
+- Add deposit request hash to contract fixture
 - Add subgraph query functionality + test script to ensure it works
 - Add EIP712 signing logic + script for generating fixture